Feature/jiwon/group2

See merge request s08-webmobile2-sub2/S08P12D110!35

+++ back_Nanal/src/main/java/com/dbd/nanal/repository/UserRepository.java
@@ -1,11 +1,12 @@
 package com.dbd.nanal.repository;
 
 import com.dbd.nanal.model.UserEntity;
+import io.swagger.models.auth.In;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
+public interface UserRepository extends JpaRepository<UserEntity, Integer> {
 
 
     UserEntity findByUserId(String userId);
